cosmetic changes, switch to png, change in contact listener
using concrete slabs makes it less likely to catch on an "edge of
unprecisioness" when sliding from one slab to another.
Still this is annoying.

png supports transparency, thats why we switch to it

changed the contact listener, works better now, but still not perfect.
it uses a vector of spacial object ids that are added and removed as the
contacts take place and are cleared when the character jumps.

We could also call a function from time to time to determine wether or
not contacts are still there...